third_party/blink/web_tests/fast/css/clear-activechain-list-shadow-dom-expected.txt

CONSOLE WARNING: line 25: Element.createShadowRoot is deprecated and will be removed in M80, around February 2020. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 and https://developers.google.com/web/updates/2019/07/web-components-time-to-upgrade for more details.
Click Me

Chain of active elements should be cleared including the Shadow DOM elements

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS background is "rgb(0, 255, 0)"
PASS background is "rgb(255, 0, 0)"
PASS background is "rgb(0, 255, 0)"
PASS successfullyParsed is true

TEST COMPLETE
